Add default template
This is a very basic template for RESTful APIs.
This project uses postgres as the main database, but can be easily
changed since it uses Sequelize to intermediate the db communication.

In config/config.json you can edit the database configuration, and on
config/db.cfg you can find the queries to initialise the database.

In models/ you can define db tables with their fields and associations.

In routes/ you can define a pipeline of functions for the requests,
consisting of Middlewares and Controllers.

In middlewares/ you can define functions that validate and complete the
request (example getting the session user from the header token, or
checking if an unique field already exists in the db). For error
handling call next() with the specific parameters, the error handling is
done in app.js.

In controllers/ you can define the operations on the database and
sending the results.

In tests/ you can write tests to verify the correctness of the
implementation over time. Mocha and Chai frameworks are used.